move error function to error.py
